Scene Switcher Added, Professor World Object Implementation Finished
Added a custom dynamic scene switcher which persists the world_scene. This can be extended to keep rooms persistent as well in an array. With grouping, this will also allow for a super code-efficient check on whether or not the current scene should be deleted. I intend to add a previous_scene var in the future.